// src/app/shared/types/article.interface.ts
import {ProfileInterface} from './profile.interface';

export interface ArticleInterface {
  slug: string;
  title: string;
  description: string;
  body: string;
  tagList: Array<string>;
  createdAt: string;
  updatedAt: string;
  favorited: boolean;
  favoritesCount: number;
  author: ProfileInterface;
}